requirements.txt
cachelib==0.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cs50==9.0.0
distlib==0.3.4
filelock==3.7.1
Flask==2.1.1
Flask-Login==0.6.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.1
zipp==3.8.0
